code/requirements.txt
accelerate==0.20.3
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.0
altair==5.0.1
annotated-types==0.5.0
anyio==3.7.1
argcomplete==3.1.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
boto3==1.28.5
botocore==1.31.5
Brotli==1.0.9
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
contourpy==1.1.0
cycler==0.11.0
dacite==1.7.0
decorator==5.1.1
Deprecated==1.2.14
diffusers==0.11.1
dill==0.3.6
dnspython==2.4.0
eventlet==0.33.3
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.100.0
ffmpy==0.3.0
fiftyone==0.21.4
fiftyone-brain==0.13.0
fiftyone-db==0.4.0
filelock==3.12.2
fonttools==4.41.0
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
glob2==0.7
gradio==3.20.1
graphql-core==3.2.3
greenlet==2.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
hypercorn==0.14.4
hyperframe==6.0.1
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
importlib-resources==6.0.0
inflate64==0.3.1
ipdb==0.13.13
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
jsonlines==3.1.0
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
kaleido==0.2.1
kiwisolver==1.4.4
lazy_loader==0.3
linkify-it-py==2.0.2
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mongoengine==0.24.2
motor==3.2.0
multidict==6.0.4
multivolumefile==0.2.3
networkx==3.1
numpy==1.25.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-ml-py==11.525.131
nvitop==1.1.2
opencv-python-headless==4.7.0.68
orjson==3.9.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
plotly==5.15.0
pprintpp==0.4.0
priority==2.0.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py7zr==0.20.5
pybcj==1.0.1
pycryptodome==3.18.0
pycryptodomex==3.18.0
pydantic==2.0.3
pydantic_core==2.3.0
pydub==0.25.1
Pygments==2.15.1
pymongo==4.4.1
pyparsing==3.0.9
pyppmd==1.0.0
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzstd==0.15.9
qudida==0.0.4
rarfile==4.0
referencing==0.29.1
regex==2023.6.3
requests==2.31.0
retrying==1.3.4
rpds-py==0.8.10
s3transfer==0.6.1
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sse-starlette==0.10.3
sseclient-py==1.7.2
stack-data==0.6.2
starlette==0.26.1
strawberry-graphql==0.138.1
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.3.0
texttable==1.6.7
threadpoolctl==3.2.0
tifffile==2023.7.10
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
traitlets==5.9.0
# Editable install with no version control (transformers==4.32.0.dev0)
-e /home/wangye/transformers-main
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uc-micro-py==1.0.2
universal-analytics-python3==1.1.1
urllib3==1.26.16
uvicorn==0.23.0
voxel51-eta==0.10.0
wcwidth==0.2.6
websockets==11.0.3
wrapt==1.15.0
wsproto==1.2.0
xmltodict==0.13.0
yarl==1.9.2
zipp==3.16.2
